// vivah-ui/src/app/profile/models/family-details.ts
export interface FamilyDetails {
    IdFamilyType: number;
    FatherName: string;
    FatherOccupation: string;
    MotherName: string;
    MotherOccupation: string;
    TotalSisters: number;
    MarriedSisters: number;
    TotalBrothers: number;
    MarriedBrothers: number;
    Address: string;
    PhoneNumber: string;
    IdPersonalDetails: number;
}